Pressing Tiles Mechanics
Added Tile Prefab
Added tile interaction script. (Tiles interact on contact with the player or other objects that have the Rigidbody component )
Tested and fully functional.

Tiles are can be modified in the Inspector.

Video explanations will be sent on Discord about how to use grabbable objects and tiles.

NOTE: on jump physics:
     In order for the character to jump the object it is standing on must have the ground Layer assigned.